feat: remove customGpts
