+++ api/models/index.js
@@ -1,5 +1,4 @@
 const { getMessages, saveMessage, deleteMessagesSince, deleteMessages } = require('./Message');
-const { getCustomGpts, updateCustomGpt, updateByLabel, deleteCustomGpts } = require('./CustomGpt');
 const { getConvoTitle, getConvo, saveConvo, updateConvo } = require('./Conversation');
 const { getPreset, getPresets, savePreset, deletePresets } = require('./Preset');
 
@@ -13,11 +12,6 @@ module.exports = {
   getConvo,
   saveConvo,
   updateConvo,
-
-  getCustomGpts,
-  updateCustomGpt,
-  updateByLabel,
-  deleteCustomGpts,
 
   getPreset,
   getPresets,
