+++ api/server/routes/index.js
@@ -2,10 +2,11 @@ const ask = require('./ask');
 const messages = require('./messages');
 const convos = require('./convos');
 const presets = require('./presets');
-const customGpts = require('./customGpts');
 const prompts = require('./prompts');
 const search = require('./search');
 const tokenizer = require('./tokenizer');
+const me = require('./me');
+const endpoints = require('./endpoints');
 const { router: auth, authenticatedOr401, authenticatedOrRedirect } = require('./auth');
 
 module.exports = {
@@ -14,10 +15,11 @@ module.exports = {
   messages,
   convos,
   presets,
-  customGpts,
   prompts,
   auth,
   tokenizer,
+  me,
+  endpoints,
   authenticatedOr401,
   authenticatedOrRedirect
 };
