+++ client/src/store/index.js
@@ -1,6 +1,5 @@
 import conversation from './conversation';
 import conversations from './conversations';
-import models from './models';
 import endpoints from './endpoints';
 import user from './user';
 import text from './text';
@@ -11,10 +10,9 @@ import preset from './preset';
 export default {
   ...conversation,
   ...conversations,
-  ...models,
   ...endpoints,
   ...user,
-  text,
+  ...text,
   ...submission,
   ...search,
   ...preset
